refactor: use copyWIth
