+++ lib/app/components/text_span_builder/text_span_builder.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:ion/app/extensions/extensions.dart';
 import 'package:ion/app/router/app_routes.c.dart';
 import 'package:ion/app/services/browser/browser.dart';
-import 'package:ion/app/services/text_parser/text_match.dart';
-import 'package:ion/app/services/text_parser/text_matcher.dart';
+import 'package:ion/app/services/text_parser/model/text_match.c.dart';
+import 'package:ion/app/services/text_parser/model/text_matcher.dart';
 
 /// Constructs TextSpan from TextMatch objects with customizable styles and tap handling.
 class TextSpanBuilder {
